Create Vote entity

+++ landmarks-api/src/main/java/dev/demo/landmarks/web/dto/LandmarkResponse.java
@@ -23,7 +23,7 @@ public class LandmarkResponse {
 
     private Importance importance;
 
-    private Float score;
+    private Double score;
 
     private Set<String> images;
 }
